fix(exp): rename to `sshutils` package name
