+++ hcloud/exp/kit/sshutils/ssh_key_test.go
@@ -1,4 +1,4 @@
-package sshkey
+package sshutils
 
 import (
 	"strings"
